+++ examples/02.rs
@@ -2,12 +2,12 @@
 
 type Rt<T = ()> = Result<T, Box<dyn std::error::Error>>;
 
-pub struct WindowBuilder<T> {
+pub struct WindowBuilder<T: Clone> {
     gui: GUI,
     data: T,
 }
 
-impl<T> WindowBuilder<T> {
+impl<T: Clone> WindowBuilder<T> {
     fn build(self) -> Rt<Window<T>> {
         Ok(Window { data: self.data })
     }
@@ -25,7 +25,7 @@ rcstruct::rcstruct! {
             {}
         }
 
-        pub fn window<T>(&self, data: T) -> WindowBuilder<T> {
+        pub fn window<T>(&self, data: T) -> WindowBuilder<T> where T: Clone {
             let gui = outer().unwrap(); // <= `outer()` returns Option<GUI>
             WindowBuilder { gui, data }
         }
@@ -34,6 +34,6 @@ rcstruct::rcstruct! {
 
 fn main() -> Rt {
     let gui = GUI::new()?;
-    let window = gui.window(()).build()?;
+    let window = gui.window(42).build()?;
     Ok(())
 }